wego/requirements.txt
Django==2.0
djangorestframework==3.7.3
psycopg2==2.7.3.2
arrow==0.12.0
requests==2.18.4
django-simple-captcha==0.5.6
pycrypto==2.6.1